+++ niuma-admin/src/main/java/com/niuma/admin/dto/EnterDTO.java
@@ -4,6 +4,7 @@ import lombok.Data;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 /**
  * 玩家进入游戏请求体
@@ -22,5 +23,5 @@ public class EnterDTO {
     /**
      * 玩家本次进入房间携带积分；为空时使用房间最低携带积分。
      */
-    private Long carryScore;
+    private BigDecimal carryScore;
 }
